requirements.txt
python-dotenv==1.0.0
Django==4.2.7
django-debug-toolbar==4.2.0
psycopg2-binary==2.9.9
mysqlclient==2.2.0
djangorestframework==3.14.0
drf-nested-routers==0.93.4
django-filter==23.3
djoser==2.2.2
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
django-cors-headers==4.3.1
django-templated-mail==1.1.1
celery==5.3.5
redis==5.0.1
pytest==7.4.3
pytest-django==4.7.0
pytest-watch==4.2.0
model-bakery==1.17.0